Move TIO pages into src/pages/Office
